requirements.txt
bluepy==1.3.0
pkg-resources==0.0.0
pycrypto==2.6.1
